requirements.txt
pytest
requests
psycopg2
python-dotenv
jsonschema